requirements.txt
smartapi-python
logzero
websocket-client
pyotp
pandas
requests
streamlit
telethon
